Update date parsing to only guess future dates
Typing “wednesday” when today is Saturday was suggesting a
past date, it should always be in the future. This is an
undocumented feature in Chrono:
wanasit/chrono#118

I can’t think of why this would happen, but if a date in the
past really needs to be entered, typing it in full works.