Add ability to copy values from reimbursements
This makes it easier for working with the ledger. It’s shamefully
lacking in automated testing, I guess it’s not possible to access
the contents of the clipboard, sensibly enough. I could extract
the ReimbursementCollection for testing but I can’t bring myself
to do it at the moment.