Change layout to be within viewport
The previous layout was causing weirdness where opening a dialogue
from an item near the bottom of a table would scroll the table upward,
making it near-unusable on mobile, and just generally disruptive. I
wasn’t following Ember Paper conventions around having the body confined
to the viewport and using overflow to scroll the children.

This closes #8 and closes #37.