Change user count to deduplicate
I had it counting duplicates for the sandbox only, really, since
on the production site, you’d generally know if you were logged
in on multiple browsers. The counts were getting ridiculous when
I left browsers open between machine sleeps, maybe because of
disconnexion and reconnexion. Rather than try to debug for this
not-really-that-useful feature, I’m just removing it.