Remove interface for logged-in report-submission
This is an unfortunate temporary solution to force report
listing/submission along the same API paths. If I could
figure out how to configure Ember Simple Auth to let
ride-fetching and submission happen without including the
token, I wouldn’t be taking this sad step.

It may be that my choice to make /rides respond differently
when an admin vs not was a mistake, maybe I should use
/rides/reportable or the like. I’ll figure something out
eventually.