Remove use of buffered proxy for reporting
This is a UX change to work around an instance of this error:
https://github.com/emberjs/ember.js/issues/16541

Previously, the entire form showed before the user had chosen
a ride; now, the user has to choose a ride before the rest of
the form shows. I don’t love changing the UX as a workaround
for this error, but I couldn’t figure out another way.