Fix ride form after errors on saving new record
This closes #123, though at the unfortunate cost that
buffered changes will be applied to the source model
after a failed save; at least people can now correct
the errors on a new record rather than having to open
a new form, but now validation errors when updating
a model will cause it to change in the background.
I believe that editing is less common, and especially
editing that will trigger validation errors, but it’s
still subpar, and I’m surprised I haven’t been able
to find any discussion of this pattern.